Move spaces and MARL utils to an internal folder
