+++ source/extensions/omni.isaac.lab/omni/isaac/lab/envs/direct_rl_env.py
@@ -31,6 +31,7 @@ from omni.isaac.lab.utils.timer import Timer
 from .common import VecEnvObs, VecEnvStepReturn
 from .direct_rl_env_cfg import DirectRLEnvCfg
 from .ui import ViewportCameraController
+from .utils.spaces import sample_space, spec_to_gym_space
 
 
 class DirectRLEnv(gym.Env):
@@ -506,9 +507,6 @@ class DirectRLEnv(gym.Env):
     """
 
     def _configure_gym_env_spaces(self):
-        # defer import to avoid circular import error
-        from omni.isaac.lab.envs.utils import sample_space, spec_to_gym_space
-
         """Configure the action and observation spaces for the Gym environment."""
         # show deprecation message and overwrite configuration
         if self.cfg.num_actions is not None:
