+++ source/extensions/omni.isaac.lab/omni/isaac/lab/envs/utils/__init__.py
@@ -1,0 +1,6 @@
+# Copyright (c) 2022-2024, The Isaac Lab Project Developers.
+# All rights reserved.
+#
+# SPDX-License-Identifier: BSD-3-Clause
+
+"""Sub-package for environment utils."""
